fix some bugs and add some contents

+++ src/Controller/AssociationBackController.php
@@ -15,12 +15,12 @@ use Dompdf\Options;
 
 #[Route('/dashboard/association')]
 class AssociationBackController extends AbstractController
-
 {
-    
     #[Route('/pdf', name: 'pdf', methods: ['GET'])]
     public function pdf(AssociationRepository $AssociationRepository): Response
     {
+        if (!$this->getUser())
+            return $this->redirectToRoute('app_login');
         // Configure Dompdf according to your needs
         $pdfOptions = new Options();
         $pdfOptions->set('defaultFont', 'Arial');
@@ -55,6 +55,9 @@ class AssociationBackController extends AbstractController
     #[Route('/', name: 'association_back_index', methods: ['GET'])]
     public function index(AssociationRepository $associationRepository): Response
     {
+        if (!$this->getUser())
+            return $this->redirectToRoute('app_login');
+
         return $this->render('back/associationback/index.html.twig', [
             'associations' => $associationRepository->findAll(),
         ]);
@@ -63,6 +66,9 @@ class AssociationBackController extends AbstractController
     #[Route('/new', name: 'association_back_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->getUser())
+            return $this->redirectToRoute('app_login');
+
         $association = new Association();
         $form = $this->createForm(AssociationType::class, $association);
         $form->handleRequest($request);
@@ -83,6 +89,9 @@ class AssociationBackController extends AbstractController
     #[Route('/{id}', name: 'association_back_show', methods: ['GET'])]
     public function show(Association $association): Response
     {
+        if (!$this->getUser())
+            return $this->redirectToRoute('app_login');
+
         return $this->render('back/associationback/show.html.twig', [
             'association' => $association,
         ]);
@@ -91,6 +100,9 @@ class AssociationBackController extends AbstractController
     #[Route('/{id}/edit', name: 'association_back_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Association $association, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->getUser())
+            return $this->redirectToRoute('app_login');
+
         $form = $this->createForm(AssociationType::class, $association);
         $form->handleRequest($request);
 
@@ -109,6 +121,9 @@ class AssociationBackController extends AbstractController
     #[Route('/{id}', name: 'association_back_delete', methods: ['POST'])]
     public function delete(Request $request, Association $association, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->getUser())
+            return $this->redirectToRoute('app_login');
+
         if ($this->isCsrfTokenValid('delete'.$association->getId(), $request->request->get('_token'))) {
             $entityManager->remove($association);
             $entityManager->flush();
